Add a base EditView class
